requirements.txt
langchain-openai
langchain
langchain_community
streamlit
python-dotenv
langserve 
fastapi
uvicorn
sse_starlette